[FE] refactor: 마이페이지 메뉴 폴더 생성 및 이동
