+++ FE/src/components/mypage/MyPageFrame.tsx
@@ -3,8 +3,8 @@ import styles from "@styles/mypage/MyPage.module.scss";
 import MyPageList from "./MyPageList";
 import MyPageCategoryItem from "./item/MyPageCategoryItem";
 import MyPageDetail from "./MyPageDetail";
-import MyPageCategory from "./details/MyPageCategory";
-import MyPageDday from "./details/MyPageDday";
+import MyPageCategory from "./details/diary/MyPageCategory";
+import MyPageDday from "./details/diary/MyPageDday";
 import MyPageDdayItem from "./item/MyPageDdayItem";
 import { categoryType, ddayType } from "@util/planner.interface";
 import { CATEGORY_LIST, CATEGORY_COLORS } from "@util/data/CategoryData";
