+++ FE/src/components/mypage/details/friend/MyPageFriend.tsx
