+++ FE/src/components/mypage/details/myInfo/MyPageInfo.tsx
